requirements.txt
streamlit
pandas
numpy
openpyxl
ofxparse
reportlab
requests
python-docx
pdfplumber
psycopg2-binary
python-dotenv
